pyproject.toml
[project]
name = "rlm-forge"
version = "0.1.0"
description = "RLM-Forge: Recursive Language Model training environment for AI coding agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "freezegun>=1.5.5",
    "gitpython>=3.1.0",
    "openenv-core[core]>=0.2.0",
    "pydantic>=2.0.0",
    "pytest>=9.0.2",
    "requests>=2.31.0",
    "text-unidecode>=1.3",
    "uvicorn[standard]>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["rlm_forge*"]
